Move iframe integration logic out of Authenticator object. This has now been implemented in TN.
